rename view_holder package to viewholders
