+++ app/src/main/java/org/stepic/droid/di/AppCoreComponent.kt
@@ -29,7 +29,7 @@ import org.stepic.droid.receivers.DownloadClickReceiver
 import org.stepic.droid.receivers.InternetConnectionEnabledReceiver
 import org.stepic.droid.services.*
 import org.stepic.droid.ui.adapters.*
-import org.stepic.droid.ui.adapters.view_hoders.CourseItemViewHolder
+import org.stepic.droid.ui.adapters.viewhoders.CourseItemViewHolder
 import org.stepic.droid.ui.custom.ExpandableTextView
 import org.stepic.droid.ui.custom.LatexSupportableEnhancedFrameLayout
 import org.stepic.droid.ui.custom.LatexSupportableWebView
